Implement command and pinning system for points, and a very unfortunate refactor to the two main pages
- Added support for command points with special icon and styling
- Introduced pinnedPointId to spacesTable for tracking pinned points
- Enhanced PointCard to display command and pin-related badges
- Updated fetchFeed, fetchPoints, and fetchPointNegations to include command and pinning metadata
- Implemented command execution logic in makePoint action
- Added isCommand column to pointsTable to track command points
- Created handleCommand action for processing command points
- Updated various query hooks to support new command and pinning features
- Query invalidation overhaul
- split big pages into client side and server side
- also fixed missing escape for ' in the pointcards
