+++ src/mutations/useDoubt.ts
@@ -4,6 +4,7 @@ import { useInvalidateRelatedPoints } from "@/queries/usePointData";
 import { userQueryKey } from "@/queries/useUser";
 import { usePrivy } from "@privy-io/react-auth";
 import { useQueryClient } from "@tanstack/react-query";
+import { pointQueryKey } from "@/queries/usePointData";
 
 export const useDoubt = () => {
   const queryClient = useQueryClient();
@@ -20,11 +21,17 @@ export const useDoubt = () => {
       // Invalidate doubt-specific queries
       queryClient.invalidateQueries({
         queryKey: ["doubt", pointId, negationId],
+        exact: false,
       });
 
-      // Invalidate point-negations to update percentages
+      // Invalidate point-negations to update relationships and icons
       queryClient.invalidateQueries({
-        queryKey: [pointId, "point-negations"],
+        queryKey: ["point-negations", pointId],
+        exact: false,
+      });
+      queryClient.invalidateQueries({
+        queryKey: ["point-negations", negationId],
+        exact: false,
       });
 
       // Invalidate favor history
@@ -45,6 +52,22 @@ export const useDoubt = () => {
       // Invalidate feed
       queryClient.invalidateQueries({
         queryKey: ["feed"],
+      });
+
+      // Invalidate pinned point in case either point is pinned
+      queryClient.invalidateQueries({
+        queryKey: ["pinnedPoint"],
+        exact: false,
+      });
+
+      // Force refetch of point data
+      queryClient.refetchQueries({
+        queryKey: pointQueryKey({ pointId, userId: user?.id }),
+        exact: true,
+      });
+      queryClient.refetchQueries({
+        queryKey: pointQueryKey({ pointId: negationId, userId: user?.id }),
+        exact: true,
       });
     },
   });
