+++ src/queries/useSpace.ts
@@ -4,10 +4,13 @@ import { useQuery } from "@tanstack/react-query";
 import { usePathname } from "next/navigation";
 import { useMemo } from "react";
 
-export const useSpace = () => {
+export const useSpace = (spaceOverride?: string) => {
   const pathName = usePathname();
 
-  const space = useMemo(() => getSpaceFromPathname(pathName), [pathName]);
+  const space = useMemo(() => {
+    // If spaceOverride is provided, use it; otherwise get from pathname
+    return spaceOverride || getSpaceFromPathname(pathName);
+  }, [pathName, spaceOverride]);
 
   return useQuery({
     queryKey: ["space", space],
